task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,17 +7,17 @@
 
 int main(void)
 {
-	unsigned long long int limit = 98;
-	unsigned long long int term1 = 1;
-	unsigned long long int term2 = 2;
-	unsigned long long int count = 2;
-	unsigned long long int next;
+	unsigned long int limit = 98;
+	unsigned long int term1 = 1;
+	unsigned long int term2 = 2;
+	unsigned long int count = 2;
+	unsigned long int next;
 
-	printf("%llu, %llu", term1, term2);
+	printf("%lu, %lu", term1, term2);
 	while (count <= limit)
 	{
 		next = term1 + term2;
-		printf(", %llu", next);
+		printf(", %lu", next);
 		term1 = term2;
 		term2 = next;
 		count++;
